+++ packages/react/src/themes/teams/components/Tooltip/tooltipContentStyles.ts
@@ -1,12 +1,16 @@
 import { ComponentSlotStylesInput, ICSSInJSStyle } from '../../../types'
-import { PopupContentProps } from '../../../../components/Popup/PopupContent'
-import { PopupContentVariables } from './popupContentVariables'
+import { TooltipContentProps } from '../../../../components/Tooltip/TooltipContent'
+import { TooltipContentVariables } from './tooltipContentVariables'
 import getPointerStyles from '../../getPointerStyles'
 
-const popupContentStyles: ComponentSlotStylesInput<PopupContentProps, PopupContentVariables> = {
+const tooltipContentStyles: ComponentSlotStylesInput<
+  TooltipContentProps,
+  TooltipContentVariables
+> = {
   root: ({ props: p, variables: v, rtl }): ICSSInJSStyle => ({
     borderRadius: v.borderRadius,
     display: 'block',
+    maxWidth: v.maxWidth,
 
     ...(p.pointing && getPointerStyles(v.pointerOffset, v.pointerMargin, rtl, p.placement).root),
   }),
@@ -15,8 +19,6 @@ const popupContentStyles: ComponentSlotStylesInput<PopupContentProps, PopupConte
     position: 'absolute',
 
     backgroundColor: 'inherit',
-    borderBottom: `${v.borderSize} solid ${v.borderColor}`,
-    borderRight: `${v.borderSize} solid ${v.borderColor}`,
 
     height: v.pointerSize,
     width: v.pointerSize,
@@ -27,10 +29,9 @@ const popupContentStyles: ComponentSlotStylesInput<PopupContentProps, PopupConte
     display: 'block',
     padding: v.padding,
 
-    border: `${v.borderSize} solid ${v.borderColor}`,
     borderRadius: 'inherit',
-    boxShadow: `0 2px 4px 0 ${v.borderColor}, 0 2px 10px 0 ${v.borderColor}`,
+    boxShadow: `${v.boxShadowStart} ${v.boxShadowColor}, ${v.boxShadowEnd} ${v.boxShadowColor}`,
   }),
 }
 
-export default popupContentStyles
+export default tooltipContentStyles